+++ simulation.c
@@ -23,8 +23,8 @@ Simulation ag_simulation_new() {
   Simulation simulation = {
       .actors = ag_actor_group_new(),
   };
-  for (int row = 0; row < SIMULATION_HEIGHT; row++) {
-    for (int col = 0; col < SIMULATION_WIDTH; col++) {
+  for (size_t row = 0; row < SIMULATION_HEIGHT; row++) {
+    for (size_t col = 0; col < SIMULATION_WIDTH; col++) {
       simulation.patches[row][col] = ag_actor_new();
     }
   }
@@ -38,8 +38,8 @@ void ag_simulation_destroy(Simulation *simulation) {
 void ag_simulation_run(Simulation *simulation, ActorUpdate actor_update,
                        ActorUpdate patch_update) {
   ag_actor_group_perform(simulation->actors, actor_update);
-  for (int row = 0; row < SIMULATION_HEIGHT; row++) {
-    for (int col = 0; col < SIMULATION_WIDTH; col++) {
+  for (size_t row = 0; row < SIMULATION_HEIGHT; row++) {
+    for (size_t col = 0; col < SIMULATION_WIDTH; col++) {
       Actor *patch = &simulation->patches[row][col];
       patch_update(patch);
     }
